[OF-1689] test: Port most multiplayer tests to test both engines
Excepting any tests in `MultiplayerTests` that are very specifically about online behaviour, take all the other tests and move them to `RealtimeEngineEntityTests`. Make all of these tests data driven and test both single and multi engines.

A couple of tests were moved to `SpaceEntityTest` as they were unit tests and it was more appropriate.

This'll be a bit of a slog, you can skim it. Almost everything is identical except blocks that do if (RealtimeEngineType == csp::common::RealtimeEngineType), so you can just focus on those ... and even then, nothing has really changed, just branching based on differing expectations.

Found several bugs doing this, which were fixed in prior commits.
